[scheduler] Move protos with trigger bodies from internal into API.
We'll add new BuildbucketTrigger proto there and then will use them in new
EmitTriggers RPC.

Keep the full-fledged Trigger (the one containing all metadata) in the internal
still, there's no need to expose it to the outside world for now.

Also drop "Data" suffix in the proto message names, it is redundant now.

R=tandrii@chromium.org, nodir@chromium.org
BUG=794499, 764043

Change-Id: Iba1508d71cacf3627cf43f01994466191d26dc85