[scheduler] Allow to order triggers by time they were emitted.
This is needed to sort triggers (which are stored as unordered sets in v2
version of the engine) when there's no better comparator available.

Be careful to handle "emitted at exact same moment in time" situation correctly
by using tuple (timestamp, order in batch).

This CL populates the fields, but doesn't use them yet.

R=tandrii@chromium.org
BUG=794499, 764043

Change-Id: I5a25955eacddc10a2bf0a0b69ddb530eaeb2dda2