[gerrit] Handle encoding of '/' correctly
If a change ID contains a '/' (consider infra/config~master~<change-id>)
then it gets encoded twice, once via url.PathEscape, and again via the
URL object. This can be avoided by assuming the path passed to get() and
post() are passed to the URL object as Opaque, which indicates that the
string has been encoded in an opaque way. Ultimately, this results in a
404 response from Gerrit if the change ID contains a '/' character.

Change-Id: I70b8642c66d1a3e60fc2e2911ae9d5ff2d34b6fb