[tsmon] Remove 'error' from tsmon metric methods.
Almost no one checks these errors, and they are never happening in a well formed
program.

Instead we now panic if the program is malformed (e.g. it tries to pass an
integer for string-typed field or "increment" a string-typed metric), or log
an error if runtime values are wrong (e.g. when counter is being rolled back).

R=iannucci@chromium.org
BUG=807844

Change-Id: I506c18d6f5ec6788e4db341b7e65b429bd9fe1bc