[vpython] Simplify and parallelize wheel installation.
Wheel installation is roughly a python 'unzip' operation. Currently we
use `pip` with lots of arguments to trick it into only installing all the
wheels from a local directory. It currently does a bunch of (unnecessary)
dependency resolution, followed by serially unzipping each wheel using
python's builtin unzip support.

This CL replaces the single pip invocation with a parallel set of direct
`wheel install` invocations. Wheel install is a much simpler operation (no
dependency resolution, etc.), and is safe to parallelize in this context
(since we've already selected a compatible set of wheels).

This yields about a 30% speedup on my '14 MacBook Pro (w/ SSD). I
expect to see a much greater speedup on machines with spinning rust
and/or Windows.

As a happy side-effect, this removes the need for the `vpython/wheel`
package, since vpython entirely outsources the job of wheel name parsing
to the python `wheel` command.

R=dnj@chromium.org, vadimsh@chromium.org

Bug: 804174
Change-Id: I65a5dfff122a17852d6ee4a210bd3ad89b59c6e4