[prpc] Simplify prpc command to no longer support CLI flags.
We'd like to make this binary available to general Chromium
usage (so, available in recipes as well as in depot_tools). To
that end, it's worthwhile to simplify the command to the smallest
API that we want to support in the long term.

The flags CLI is occasionally useful, but generally confusing
(especially for complex usages). The concern here is that by
deploying the prpc CLI tool, we may inadvertently have someone
(bots, users, user scripts) take a dependency on this format.

The alternative formats (JSONPB, Binary, Text) are all
well-defined and generally supported formats outside this tool.
Additionally, we've found that 9/10 times we want to use JSONPB
anyway. This CL makes JSONPB the default format.

Bug: 808677
Change-Id: I04b4ea0bbb6327f2f75cf0ef16694d7225ede54d
Reviewed-on: https://chromium-review.googlesource.com/900028
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>