Update patch set 1
Patch Set 1: Code-Review-1

This CL didn't change the way that vpython writes to the filesystem, which is the current operating theory in that bug.

Moreover, reverting this won't cause it to get deployed anywhere. If you want to try deploying an earlier version, we can certainly do that experiment. I assume that someone working on this has a local repro? I could work with them to introduce a different version of vpython to see if that behaves differently.

This particular CL has been deployed for a couple weeks at this point (and has been deployed roughly everywhere in that time period). Several CLs in chromium have landed that depend on this change. Even if reverting this would deploy it (which it won't), it would cause other builders to break.

Patch-set: 1
Reviewer: Robbie Iannucci <1001907@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=-1