[milo] Remove blank Changes column from buildbucket builder pages.
We can add it back if we determine that there's a need for this (and
a meaningful implementation).

In Buildbot, changes are directly associated with Build objects. Retries
and forced builds have no 'changes' associated with them. In
Buildbucket, builds record the fact of "I built revision X", and then
milo dynamically computes the blamelist between a build and the previous
build for that builder. This is done in order to support
backfilling/bisection on main waterfall builders at a later time.
However, due to this design, meaningful display of a changelist on this
view is more difficult than it is for buildbot.

Coupled with the fact that the primary use case for this information is
also less relevant than it was for Buildbot (namely; the implication
that the changelist here correlates with the incremental build that the
bot did. On Swarming, there can be multiple bots in the pool, the cached
build can be removed between tasks, force/retry builds won't be in
commit order, etc.)

Diagnosing this sort of thing on the bot (speed of incremental builds)
will be better served by reporting the 'currently checked out version'
from the bot before the build. Then we'll know the actual starting
revision, the checked out revision, and the duration of the build.

R=hinoka@chromium.org, jchinlee@chromium.org

Bug: 807846
Change-Id: I04727eb8487653a01deadd3e1b863082a2b863ef