Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > why not generic code that does a bit of reflection? 
> > > it is really not that hard:
> > > https://chromium.googlesource.com/infra/luci/luci-go/+/9c5cd53c3145bd8cc9bc3108669b75fc9819f26b/common/bq/eventupload.go#140
> > 
> > When I brought it up at today's meeting, Eric suggested this approach would be fine and reflection might be overkill.
> 
> Erik, I wonder if you change your mind with more context

(Sorry for confusing Erik with Eric, just confirming that I did mean Erik.)

Nodir, I think that attempting to genericize will lead us to the same place as it did much earlier when I was just starting the server-side code and you suggested reflection. We thought about it for several days and in the end you seemed to agree that the non-generic approach was best because of the similar-but-slightly-different nature of each entity which promised to complicate the usage of reflection. I'm worried about going down the same road here. Although I was originally going to use it here, after looking into it I think it will be simpler not to reflect. With reflection we need to devise a one-size-fits-all solution for printing all protos rather than tailoring it to each specific RPC. Consider proto fields, repeated proto fields, or repeated proto fields which each have their own repeated proto fields. We would need to decide that all protos will be recursed DFS or BFS to generate column order. If DFS we'll need to pick pre-, in-, or post-order to order the columns. A traversal order which is natural for one type of response may not be for another (not a hypothetical, I do plan to add repeated NICs to each Machine so machines may return all associated network interfaces in one RPC). We would also need to devise a naming scheme to differentiate between same-named fields in different protos. I believe that'd get complex quickly. Instead we could just do what makes the most sense for humans when emitting each individual response.

Patch-set: 2
Reviewer: smut <1003211@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0