[butler] Fix flaky test.
The one that results in:
 Line 442:
  Expected: 'map[string]*logpb.LogStreamDescriptor{"stdout": ...}'
  Actual:   'map[string]*logpb.LogStreamDescriptor(nil)'
  (Should resemble)!

There's a race condition there. The bundler has a chance to "drain" and
unregister &bytes.Buffer{} stream right after AddStream, before GetStreamDescs
is called. To prevent this, we make test stream's Read(...) block for the
entire duration of the test.

R=iannucci@chromium.org
BUG=762635

Change-Id: I5777a82636003009f9362bf50057463d2d50c6de
Reviewed-on: https://chromium-review.googlesource.com/918301
Reviewed-by: Daniel Jacques <dnj@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>