gkelogger - A simple JSON logger for GKE
This outputs logs to stdout in a format that the GKE (fluentd) + Stackdriver
stack expects.  The format is one json object per line.

This also replaces the default logger for logdog GKE modules to use
this logger.

Not in scope of this CL is to replace logging with per-request
(ie. pubsub message) nested logs using cloud logging.  That will be
in the future.

Bug: 812474
Change-Id: I8d4f679cf00abbe8a5fcd38d6f496b86b6cd19bd
Reviewed-on: https://chromium-review.googlesource.com/920660
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Ryan Tseng <hinoka@chromium.org>