[swarming] Return in-progress results on timeout
This change updates the swarming client to keep track of the last
retrieved set of swarming task information so that it may be returned in
the event of a timeout in the swarming client. Note that this is not the
timeout given by swarming, but rather the timeout for the swarming
client's polling itself. This will allow one to propagate information
about what the task was doing at the point where the client timed out.

Change-Id: I5229928aebc83b37eab2e0fedbb3a09543fe556a