Update patch set 2
Patch Set 2:

It actually might be better to specify "timeout" as a sort of "I got errors consistently from swarming for X minutes so give up" instead of "how long should I wait for". The reason for this is the swarming task itself already has a timeout, and the purpose of the timeout in the swarming client is more oriented toward dealing with faults in talking to swarming. If we're getting a valid result back, I think we should probably keep waiting.

WDYT?

Patch-set: 2
Reviewer: Michael Knyszek <1154295@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>