[luci-notify] Allow builds without a revision
This change enables builders without a revision to be tracked by
luci-notify for on-change notifications by explicitly storing an
empty revision in such cases, and falling back on build creation
time for ordering. Additionally, builders may be "upgraded" to being
tracked by a revision or "downgraded" to being tracked without a
revision (that is, newest build always wins).

Change-Id: Icfe6e6ee4b65de91b3d0905de775e8bb545624e8