Update patch set 2
Patch Set 2:

this CL fixes a bug in a code that should be deleted sooner rather than later, and a comment mentions another upcoming fix. Since you seem to have time for these bugs, how about fixing the underlying problem: crbug.com/774345 ?

BTW the bug that this CL fixes does not exist in buildbucket, so we wouldn't have this problem if we trusted buildbucket in the first place.

Patch-set: 2
CC: Erik Staab <1128941@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
CC: Nodir Turakulov <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>