luci-notify: Allow emails to chromium.org.
Tryjob notifications are usually sent to the accounts generating CLs,
which are usually chromium.org based.

Allow email to be sent to @chromium.org accounts, in addition to
@google.com accounts. This is a loosening of restrictions, and so
a potential loosening of security (but only if people can edit the
luci-notify configurations).

This affects luci-notify configurations, as well as any buildbucket
property based notifications when they are implemented.

BUG=chromium:816652

Change-Id: I9a05ff9b7eebdeb99581faad9d0ec1947b032591
Reviewed-on: https://chromium-review.googlesource.com/947709
Commit-Queue: Don Garrett <dgarrett@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>