[luci_notify] Change email template for better threading
The old template would cause the email thread to always show the
original status (e.g., [Build Failure]) in gmail even if the status
eventually changed to Success.

Instead, avoid mentioning the status in the subject line.

Also, make the current status stand out more in the email body, and
include the time that the status change happened.

Change-Id: I792240f439d585a13c5a09e34deb46fc7ad8b6df
Fuchsia-Bug: INTK-160