Update patch set 1
Patch Set 1: Code-Review+1

(1 comment)

LGTM and adding nodir as well to ChOps is in the loop.

As long as we're doing this, it might also be worth inserting some red/green/purple coloring into the email (it is just HTML after all), but that's up to you.

Thanks for doing this, btw!

Patch-set: 1
Reviewer: Michael Knyszek <1154295@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1