luci-notify: Populate nil referernce to buildbucket.Build.
A pointer to an inner struct wasn't being populated. Test code uses a
mockup helper that does populate it, and so wasn't seeing the
same panic.

This was missed because there is no test coverage for extractBuild,
but adding that coverage requires creating populated http.Request
objects, which is a larger project than I want to address today.

BUG=chromium:816652

Change-Id: Ib00bd025fa7a829a84425c0f15fdf7ba39e059e6