Revert "[common/proto] add monitor and step_class to step"
This reverts commit d09e84625b2a683a99641b278c612e7c8adf3c7c.

Reason for revert: step class concept wasn't integrated into anything and now we are short on staffing for this project, so it is better not to have a half-done implementation. This CL just adds a field, we can restore it at any time.

Also, I think luci-eng@ should converge on step class concept before introducing it.

Original change's description:
> [common/proto] add monitor and step_class to step
> 
> Bug: 795445
> Change-Id: Id046214d8129dd48c8e49070693b782c7f09485b
> Reviewed-on: https://chromium-review.googlesource.com/944517
> Commit-Queue: Katie Thomas <katthomas@google.com>
> Reviewed-by: Robbie Iannucci <iannucci@chromium.org>

TBR=iannucci@chromium.org,nodir@chromium.org,katthomas@google.com

# Not skipping CQ checks because original CL landed > 1 day ago.

Bug: 795445
Change-Id: I7dc74745e12ffe3ecbc4311b15efa77b522a5b7f