[cipd] Add RPC definitions for prefix metadata RPCs.
Prefix metadata is a new concept that generalizes path package ACLs.

There's a need to attach various structured information to directory nodes in
CIPD package namespace (aka "prefixes"). Things like:
  1. Currently supported ACLs (reader/writer/owner).
  2. Refs/tags ACLs ("can update tags with key 'git_revision'").
  3. Name of the canonical tag ("first unique tags with 'git_revision' key
     should become the canonical version name).
  4. "Do not inherit ACLs" flag.
  5. A contact email ("poke the given group if you want to get access").
  6. Package namespace description ("Packages here contain blah", for UI).
  ...

The current datastore structure used for ACLs is very rigid and insufficient to
support these use cases. We'll refactor it (including migrating existing data,
which is tiny) to support storing PrefixMetadata protos as they are, with all
their rich extendable structure.

RPCs in this CL define the new interface to deal with them.

R=nodir@chromium.org, iannucci@chromium.org
BUG=801423

Change-Id: I7edbe954bc8bdfaf149578c215d214821db2dfb9