Update patch set 6
Patch Set 6:

Added tests.

I've also realized it is OK to return NotFound for non-existing prefixes that are under owned prefix, i.e. if the caller is OWNER of prefix "a", and "a/b" doesn't exist:

GetPrefixMetadata("a/b") => NotFound
GetPrefixMetadata("some-other-prefix") => PermissionDenied

In this example "some-other-prefix" may or may not exist, but in any case it is not visible by the caller, so callers gets PermissionDenined. See tests for more cases.

(I'll update the documentation in the proto in https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/974881).

Patch-set: 6
Reviewer: Vadim Shtayura <1002539@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>