Create configuration for luci-notify email templates.
This creates the concept of email templates in the luci-notify
configuration files, but doesn't do anything with them yet.

An arbitrary number of named templates can be defined and used by name
from both build requests in json_parameters, and from notifier
defitions.

The first template defined will be used as the default for all project
notifications not otherwise specified. The existing hard coded
template will be used for all projects with no template definitions.

Bug: chromium:816652
Change-Id: I7818334943e593d98cd8322f2c09498a44cb74cd