Create configuration for luci-notify email templates.
This creates the concept of email templates in the luci-notify
configuration files, but doesn't do anything with them yet.

An arbitrary number of named templates can be defined and used by name
from both build requests in parameters_json, and from notifier
defitions.

The first template defined will be used as the default for all project
notifications not otherwise specified.

Bug: chromium:816652
Change-Id: I5f6d1320358b6d4c6ec97f1576c9b4f8529fb60a