[Machine Database] Replace internalError with grpcutil.GRPCifyAndLogErr
Before, each function had to worry about how its errors might be used and call
internalError() to obfuscate them if they might ever be seen by an end user, the
result of which was that everything down the call stack would not have access to
the actual error. Now allow the call stack to see the full error, and instead
defer the obfuscation decision to right before the user sees it.

Note that pRPC automatically obfuscates errors with codes.Internal or
codes.Unknown with the message "Internal Server Error", and GRPCifyAndLogErr
converts all non-gRPC errors into Unknown, and therefore obfuscated, gRPC
errors.

Bug: 778865
Change-Id: Iaad8902bcf452a148a8bff44312fdd2b7b7c2b5c