Plumb template names from config to generation.
Now that we can configure template names to use when sending email, we
need to keep those template names associated with the email addresses
that should get the email.

This CL creates that plumbing, and sends an email per template, but
does not actually parse or use the templates themselves.

Not yet ready.

BUG=chromium:816652

Change-Id: Iafa35c24f0080ec20ff43b2b8d9a7114e3b02ba7