[milo] Use Buildbucket for buildbucket build pages
This uses buildbucket data to render buildbucket pages, instead of swarming data.

This also refactors things to make it easier to delete the swarming rendering code.

The only user visible change should be that the "Steps" step is removed.
This is because the "Steps" step is a special step that represents the overall summary
of the build as seen by the annotation in LogDog, not Buildbucket.
Also the Swarming bot is missing in the buildbucket data, and is TODO to be added back.

Bug: 774345
Change-Id: I5ab46ebf9d13efae8315d5296b625b5a8deadf0d
Reviewed-on: https://chromium-review.googlesource.com/976943
Commit-Queue: Ryan Tseng <hinoka@chromium.org>
Reviewed-by: Nodir Turakulov <nodir@chromium.org>