[scheduler] Implement basic cron scheduling in v2.
Each job now owns a cron state machine which emits CronTriggers for this job
based on the configured schedule. This trigger is then triaged as usual.

To support "with 10min interval" schedules, the cron machine is being rewound
during the triage if there's no invocations running anymore.

Things not supported yet:
  * Overrun status is not displayed. In fact, the new behavior is "if a job is
    running at the time new invocation should be starting, queue up the new
    invocation and launch it as soon as the previous one finishes" (this is just
    default behavior of the triggers queue).
  * Job status in UI is not displayed correctly, including "Next run" field.

This is largely based on the old prototype:
https://github.com/luci/luci-go/blob/1b403a7ea/scheduler/appengine/engine/cron/demo/main.go

R=tandrii@chromium.org
BUG=764043

Change-Id: Icc48936917e58fb9aacb1e5a4150f0deaec65444