Update patch set 3
Patch Set 3: Code-Review+1

I am a bit sad that we are adding a hack on top of another hack. The first one is that a builder group is implicitly defined by defining a console. The second one is that a group defined using a console is also marked as "actually, this is not a console. This is just a group".

In the light of the total amount of hacks in Milo, this CL doesn't make the situation a lot worse. It does, though, make me more and more convinced that we should rewrite Milo at some point, probably sometime after build api v2 is available.

Patch-set: 3
Reviewer: Nodir Turakulov <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1