Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > Looks like this broke displaying internal V8 builders.
> > Version before this CL: http://shortn/_wUMjE9nMZi.
> > Version after this CL: http://shortn/_AvuQk6h9IA.
> > Direct links to builds still work, e.g. http://shortn/_40uQMU6ync.
> 
> P.S. I've accessed all of these links being logged in with my @google.com account.

That behavior is strange to me given that this code returns a 404 if you fail to authenticate with the access API, not a 403. The 403 seems to be coming from the project check. I suspect this ultimately has to do with this TODO here? https://chromium.googlesource.com/infra/luci/luci-go/+/master/milo/frontend/routes.go#140

Patch-set: 3