[logdog] Plaintext Endpoint on Flex
This CL implements the bare minimum feature set of a plaintext endpoint.
The rest of the features (html, auth redirection, etc) will come later.

This doesn't use the fetcher/renderer paradigm because:
* It requires copying a lot of code from the existing log server renderer anyways.
* We want more control on how to render text lines.  In theory we could
augment renderer with a "WriteLine" interface that takes a LogLine and
outputs a string.  But at that point, it feels like the layer of abstraction
is causing more code to be written, rather than less.

Change-Id: I2a5b8c80c176ca59a08093ddfd2128a1b6ee0fa5