[scheduler] Implement v1 -> v2 conversion.
We introduce a new terminal state to v1 state machine called CONVERTING_TO_V2.
The v1 state machine switches to it when it finds itself in-between invocations.
While doing so, it emits ConvertToV2Action that finishes the transition to v2,
rescheduling the cron tick with v2 engine and transferring the pending triggers.

To avoid waiting for rarely changing jobs (e.g. crons that tick once a week),
we also proactively nudge v1 state machines in UpdateProjectJobs if they appear
in-between invocations.

No unit tests, since this code is going to be deleted ASAP, and it was manually
tested on luci-scheduler-dev under various conditions.

R=tandrii@chromium.org
BUG=764043

Change-Id: Iff534893a602c73fbdf100c5b5908b08c8bb7b06
Reviewed-on: https://chromium-review.googlesource.com/1029614
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Andrii Shyshkalov <tandrii@chromium.org>