Update patch set 11
Patch Set 11:

> Patch Set 10:
> 
> I'm curious why this is needed. Wouldn't it be simpler and more manageable (== explicit) if root luci-notify.cfg just references template files by a path relative to the config set root?
> 
> I'm concerned that luci-config go client is already (unfortunately and unnecessarily) complicated, and adding new methods to it should have a good justification.
> 
> In particular, you are adding ListFiles only to the lowest layer on luci-config client (basically the RPC client). There are mountains of code under https://github.com/luci/luci-go/tree/master/config/server/cfgclient and https://github.com/luci/luci-go/tree/master/config/appengine that implement datastore+local caching layer for configs (that is strongly encouraged to be used from GAE instead of hitting luci-config service all the time directly). This mountain of code would have to be updated to support ListFiles too.

Hi Vadim,

This approach was based on past discussions with Nodir regarding this initiative.  He can probably provide a bit more insight as needed, but I believe the intention is to limit the amount of config management within luci-notify itself, as the majority of all config is currently done within the luci-config client.  I can set up a side conversation if we'd like to discuss further.

In regards to the other code, the functionality exists today with the use of include_files, within the client, but is not exposed at the lower levels for use outside of the client.  Would it be necessary/should it be added at the other levels if it exists in another fashion today?  

Thanks,
Mike

Patch-set: 11
Reviewer: Mike Nichols <1273501@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>