Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> > Patch Set 10:
> > 
> > I'm curious why this is needed. Wouldn't it be simpler and more manageable (== explicit) if root luci-notify.cfg just references template files by a path relative to the config set root?
> > 
> > I'm concerned that luci-config go client is already (unfortunately and unnecessarily) complicated, and adding new methods to it should have a good justification.
> > 
> > In particular, you are adding ListFiles only to the lowest layer on luci-config client (basically the RPC client). There are mountains of code under https://github.com/luci/luci-go/tree/master/config/server/cfgclient and https://github.com/luci/luci-go/tree/master/config/appengine that implement datastore+local caching layer for configs (that is strongly encouraged to be used from GAE instead of hitting luci-config service all the time directly). This mountain of code would have to be updated to support ListFiles too.
> 
> Hi Vadim,
> 
> This approach was based on past discussions with Nodir regarding this initiative.  He can probably provide a bit more insight as needed, but I believe the intention is to limit the amount of config management within luci-notify itself, as the majority of all config is currently done within the luci-config client.  I can set up a side conversation if we'd like to discuss further.
> 
> In regards to the other code, the functionality exists today with the use of include_files, within the client, but is not exposed at the lower levels for use outside of the client.  Would it be necessary/should it be added at the other levels if it exists in another fashion today?  
> 
> Thanks,
> Mike

i think the mount of code that we have today is a mistake and we should not limit ourselves because of the mistake, if possible (possible in this case). There is a recommendation not to hit luci-config in serving code path, but it does not necessarily imply using that mount of code. Milo and luci-notify talk to luci-config in a cron job and save configs into the datastore in the format most suitable for the configured entities. For example, luci-notify indexes notifiers in a different way and would not use the mount of code. This new feature would be used only in that cron job, too. 

I considered listing all template file name in a config and thought it would be silly to ask users to add a file AND ask to specify the file name in another file. If I was such user, I'd immediately think "can't you just list the files that I provide?".

FWIW, I think we should drastically simplify the config API that we have today. Potentially even removing the file system implementation of the config.Interface

Patch-set: 11
Reviewer: Nodir Turakulov <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>