Update patch set 12
Patch Set 12:

see go/logdog-plaintext for context.

This differs from dnj@'s implementation mostly in that it doesn't try to use the fetcher/renderer paradigm, and just pretends to be a logs endpoint client. 
I think that results in less/simpler code, and is easier to add injected HTML later.

The renderer is inlined in HttpHandler().  The intention is to extract it out later
when there is more code.

Patch-set: 12