Update patch set 6
Patch Set 6:

> Patch Set 2:
> 
> In general it looks okay, but it looks like there are some bugs with sparser consoles:
> 
> https://2981-1e04250-tainted-sergiyb-dot-luci-milo-dev.appspot.com/p/v8/g/side-by-side-chromium/console

I've fixed that issue by setting flex:0 on the commits colimn and also added some logic in Go-code to calculate how wide a given category has to be in comparison with other categories. It seems to be not possible to achieve the same with flex-layout in CSS given that categories can be nested many times and that flex layout only works in parent-child and in not parent-grand..grandchild relationship.

Patch-set: 6
Reviewer: Sergiy Byelozyorov <1153089@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Commit-Queue=0