Update patch set 9
Patch Set 9:

(2 comments)

I added caching logic and changed API a bit. Since this is a significant change, PTAL. OTH, I would really like to land this ASAP, hence if you won't find time for this today, I'll land and deploy it as is and instead address your feedback in a follow-up.

Patch-set: 9
Label: Commit-Queue=0