Update patch set 1
Patch Set 1:

(4 comments)

This is fine if we don't want to go the annoying Polymer route. Some minor comments; I understand you don't want to be involved and this CL just serves as an explanation for Bev, so you decide if you want to follow up or not. If not, Bev can take over and send the CLs directly to me.

Patch-set: 1
Reviewer: smut <1003211@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>