[skylark] Add awareness of proto2 messages.
For some weird reason proto lib doesn't expose any metadata regardless whether
a message was defined as proto2 or proto3. So instead we detect unsupported
proto2 fields on the fly and fail with a clean error message (instead of
panicing as was happening before).

Such late detection has a benefit of allowing to convert stuff to proto3
gradually (i.e. it is fine to have proto2 fields in proto3 message as long as
skylark code doesn't touch their scalar fields).

R=smut@google.com
BUG=833946

Change-Id: I6b5e32fd497787ff5a624851500f04c21f4ef5bd