Update patch set 2
Patch Set 2:

Stepping back a little, this is what the configs will look like:

GooglesourceAcl {
  host: "chromium"
  project: "chromium/src"
  project: "v8/v8"
}

I think as a non-git admin, it's a little hard to understand.  Chromium is in there, but it's a host, but also a project? Whats the difference?  I'd suggest probably using URLs, which are fairly universal and difficult to misinterpret.  For example:
 
GooglesourceAcl {
  host: "https://fuchsia.googlesource.com/
  project: "https://chromium.googlesource.com/chromium/src"
  project: "https://chromium.googlesource.com/v8/v8"
}

Patch-set: 2
Reviewer: Ryan Tseng <1002080@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>