Update patch set 10
Patch Set 10: Code-Review+1

(11 comments)

I have to be honest, I am still not super happy with the code in this CL. In most happy cases its behavior is what we want, but in some error cases it does something we don't want it to do. It will cause user frustration and bug reports that nobody will have time to address because owners moved on to other work in LUCI. However, we need a more effective way to get this done. Let's land this and I will send followup CLs to fix the problems I still see.

Followup CLs are needed anyway because if deployed as is, this will start querying luci-config for each build that needs a notification, which is not scalable. I will address this too.

Patch-set: 10
Reviewer: Nodir Turakulov <1114299@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Code-Review=+1