Update patch set 10
Patch Set 10:

(7 comments)

> Patch Set 10: Code-Review+1
> 
> (11 comments)
> 
> I have to be honest, I am still not super happy with the code in this CL. In most happy cases its behavior is what we want, but in some error cases it does something we don't want it to do. It will cause user frustration and bug reports that nobody will have time to address because owners moved on to other work in LUCI. However, we need a more effective way to get this done. Let's land this and I will send followup CLs to fix the problems I still see.
> 
> Followup CLs are needed anyway because if deployed as is, this will start querying luci-config for each build that needs a notification, which is not scalable. I will address this too.

Thanks for being honest, Nodir.  

I have to be honest as well, I'm not super happy with the code.  I was frustrated by the review approach and allowed that to reflect in code being sent for review.  I apologize for the sloppiness.  I definitely own that.

With that in mind, I am also not happy with the approach in general.  Requirements were from an extremely high level "green" doc with expectations that were never communicated other than in the review.  I'm not sure if you've been on the other side, but design by review is not enjoyable by any party involved.   The requirements laid out in the review are definitely not in the design doc.  An iterative approach to these requirements would have been more efficient from both sides.  Lesson learned, at least from my part. 

Unless the expectation is that all features will be landed by the LUCI development team, more guidance and details, in the design docs and conversations, would be beneficial and can help to further the knowledge outside just the immediate LUCI team.  I gained a lot of insight and knowledge throughout the process; less frustrations and being the cause of those frustrations would have been a better outcome for all.  

Thanks again for guiding the way.

Patch-set: 10