Update patch set 1
Patch Set 1:

A bit unfortunate...

But I had to explain to someone earlier today why there is a pointy edge in our system called "source manifest", and it boiled down to "feature that nobody asked us for, and isn't currently used".  There's a time and place for this, and at the moment it is "in the future".

Add bug: 758007 also

Patch-set: 1
Reviewer: Ryan Tseng <1002080@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>