[luci-notify] Replicate email templates in the datastore
Code that runs frequently must not contact luci-config directly because
luci-config is not designed for this scale and it is slower.

In the existing cron job that fetches project config from luci-config, use
new ListFiles API to fetch names of all email templates of the project,
then fetch them and save in a new EmailTemplate datastore entity.
It happens in the same Datastore transaction as the luci-notify.cfg file.
The root entity is the project itself.

Bug: 816652
Change-Id: Ie8e856cd9a2bb67de7a2e7855e4a9bd4c5722c84