[luci-notify] add EmailTemplateBundle type
EmailTemplateBundle is a collection of all project email templates compiled
together. Goal: template in one file reuses a template in another file,
for example we have one template that defines step rendering and a few other
files that reuse the same step presentation.

Also it deletes ParsedEmailTemplate. It turned out to be not useful.

Bug: 816652

Change-Id: Iae10a894a6a7353e61f52f3c86f72b6a7d166e1b