[skylark] Add 'mutable' helper function.
It allows modules to keep a mutable internal state. Unlike Python (where module
dict is always mutable), Skylark freezes module dicts after module loading. With
'mutable' a module can get a handle to a value that doesn't get frozen.

This would allow to prototype config generation mostly entirely in Skylark. Once
we get to a final state we can move all mutable state into the Go part of the
interpreter (this is how other Skylark embedders deal with it).

R=smut@google.com
BUG=833946

Change-Id: Iff2ee6eb8ca6c8977841f13e90b94f8b9b112309