[skylark] Implement an interpreter that can load modular scripts.
Raw skylark.ExecFile and skylark.Thread facilities are limited to working with
only one file. This is not sufficient for any non-trivial scripts.

This CL introduces Interpreter struct that knows how work with a tree of skylark
scripts that import each other, as well as protobuf definition files.

It also has a support for a separate tree of script that constitute a "standard
library", with an entry point being in init.sky. All scripts executed by the
interpreter have access to symbols exported by init.sky. The standard library
can be compiled into the interpreter executable, to make it completely
standalone (this will be used in next CLs).

R=smut@google.com
BUG=833946

Change-Id: I1d2b1a8d622abcf14c0f6eaf82111c7058f29ec4
Reviewed-on: https://chromium-review.googlesource.com/1060698
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: smut <smut@google.com>