[luci_notify] Lookup only one notifier for a build
Since builders are scoped to projects, and we validate that there can
only be one copy of a builder per configuration, then logically a
builder can only belong to a single notifier. This means that wherever
we use "notifiers" in the code, that really just refers to one notifier
every single time. This change cleans this up, and uses only a single
notifier instead of a list of notifiers (whose length is always 1).