archiver: move tar archiver from cmd/isolate
That's where it should have been in the first place.

Do as little mutation as possible to keep the move reviewable.

That makes package archiver a sad package, with incompatible duplicate
implementations. :/ But it's still better than keeping one of the implementation
in a main package. A follow up unexports a few symbols.

R=mknyszek@google.com

Bug: 846712
Change-Id: I14f947fd9ea9a87fa8788296a7009fac131c4422