rebase
archiver: keep a map of seen files to skip redundant ones

During 'batch_archive', the same file may be mapped in multiple input lists
(isolate), resulting the file to be archived multiple time.

Skip duplicates at the directory walk, otherwise this causes a lot of redundant
work and may cause the same file to be stored multiple times in a tar file,
causing problems when unpacking the tarfile.

Bug: 846712
Change-Id: Ic6ff65b07930cfc6a72e8addf93b17111672b5bc