stringset: switch Set to a concrete type
There had been a plan to make a concurrent safe Set implementation but it's not
happening. Replace the interface Set with the concrete type set, which turns the
method through an interface to direct calls, which can be inlined by the
compiler.

Because of https://github.com/golang/go/issues/19348, inline all the functions
(it doesn't make the code longer!) to ensure that most methods can be inlined.

Include small performance tweaks.

Change-Id: Ib8dd36a16ab57c1383edd236ad62b78b4a10091c