[buildbucket] Allow searching for ended builds
SearchBuilds RPC does support finding completed builds regardless of build result.
This is a regression from v1.

Expand Status_COMPLETED_STATUS_MASK to represent a union of all terminal
statuses and allow it in BuildPredicate.
Rename it to ENDED. Use "end" as opposed to "complete" because it is
the antonym of "start".

Change-Id: Iad3bfcd99a50427bd204e1f08bd4569a66451f94